Add ApiClient class
The ApiClient class represents a given API site indexed by a key.
The key can be platform-dependent. On WordPress.com, for instance,
this would be a numeric siteId. On a normal wp-admin setup, it
would be the API of the site. For a Stripe API, it could be
production or test.