Change requirements to component-based mapping.
This is the first set of changes to move the primary method of gathering
requirements to a component-based approach which accounts for components
mounting and unmounting. It also switches the endpoints to a tree
structure to allow for nested endpoints. It also accounts for queries
for a given endpoint and handling requirements for those separately.