apiClient: Adjust operations to single function
The operations were originally conceived as an array of functions that
act like reducers. However, this is a bit of unneeded complexity at this
level. If pass-through functions like reducers are needed, they can be
easily implemented in the API perhaps with the aid of a helper function
like `combineOperationHandlers` or the like.