Move data handling to ApiClient
This moves the data handling and timestamp handling to the ApiClient
class, which allows this critical code to be removed from the
react-redux implementation and into the core ApiClient class. This makes
the code more standalone and not as reliant on the react-redux
implementation.

It also adds some logging for unhandled exceptions that may occur within
api operations.