Move static api members to instances
This moves `methods`, `operations`, `mutations`, and `selectors` from
static class members to instance members passed via the base class
constructor.

It will allow the `this` context to be used for these functions, and it
will also allow different `methods` to be passed in for different
instances of apis, which will prove valuable when running in different
environments.