Update version to 0.2.0
This updates the version of the `@fresh-data/framework` to 0.2.0

Changes affecting apis using `fresh-data`:

1. Instead of `( getData, requireData )`, selectors are now given
`( getResource, requireResource )`. Both of these functions return the
state of the resource by its given key.
  a. `getResource( resourceName )` returns resource state in the format
of `{ data, error, lastReceived, lastRequested, lastError }`. All
properties of that state are optional and may not be present, but
`getResource` will always return an object, even if it's empty.
  b. `requireResource( requirement, resourceName )` sets a component
requirement for the given resource, and then returns resource state in
the same manner as `getResource`.

2. mutations now return proper promises, which can be used to handle
success and failure callbacks via a Promise `resolve/reject`, or by
using `async/await`.

3. Within the client class, `methods`, `operations`, `mutations`, and
`selectors` are no longer static members, but members of the object
instance. This provides greater flexibility in API implementation by
allowing things such as passing `methods` in via a constructor.