Provider: Single api only.
This is the first step in simplifying fresh-data for 0.3.0 and getting
it to line up with the theory behind @wordpress/data a bit more.

It changes the provider to only support one api instead of a dictionary
of apis with their respective clients. No functionality is lost because
multiple providers can be rendered if multiple apis are needed to be
used.