license: Change to MIT
Since this is a general-purposes JS module on npm, it makes sense to
license it as MIT instead of the default GPL for most Automattic
projects.

"While Automattic has a strong philosophical preference for the GPL,
there may be a practical cases where it’s not possible or doesn’t
make sense to use it. Compatibility with existing projects in a field
or ecosystem might be one (e.g. projects on npm will prefer MIT).
Licensing requirements based on dependencies might be another."