Update deps for packages and examples (#160)
This updates the dependencies for the packages and examples.
It also sets the examples to depend on a specific version number instead
of whatever is in the checked-out repository, which helps decouple the
examples a little from the packages. This allows them to be updated in
subsequent PRs as needed after updating the packages themselves.