Move dataRequested/dataReceived to scheduler
This moves the data handlers from the client to the scheduler where it
will actually be used. The client class then just passes through the
data handlers when they are set to the current instance of the
scheduler.

note: this change was moved from PR #205 to PR #204 for completeness of
the scheduler implementation.