Handle operation promise order in scheduler
The previous code was causing problems with the react-provider HOC which
was evident in the examples. The client state wasn't updating until
after components would re-render, which was causing problems with the
order of updating the requirements of the render.

This forces the order of state updates and resource-request statuses by
handling the promise order explicitly in the scheduler.