Fix resources fetching twice (#234)
* Fix resources fetching twice

This fixes the problem where resources were getting fetched twice
because of the way react-redux connect works with rendering components
and sending out redux store subscriptions.

* Fix tests for setDataHandlers

The previous tests was checking for function equality. This one just
verifies that the functions are called.

* Change isStateStale to isClientStateInSync

In a package such as fresh-data, the word "stale" can be overloaded and
ambiguous. This renames it to be more specific.

* Fix requireResource tests

There were two tests where `setState` needed to be called.